+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 ## 1.0.0
 
 - Initial version.
+- Implemented brick view
+- README, LECENSE and example added.